[REF] web: split kanban test suite
This commit extracts into their own suites tests about the record
quick create and the progressbar features out of the main kanban
test suites.
